+++ main.tf
@@ -217,6 +217,57 @@ resource "aws_lb_listener_rule" "this" {
 
   tags = local.tags
 }
+
+/* ------------------------------ Green Listener Rule ----------------------------- */
+resource "aws_lb_listener_rule" "green" {
+  count = local.is_create_target_group && var.is_enable_blue_green_deployment ? 1 : 0
+
+  listener_arn = var.alb_listener_arn
+  priority     = var.alb_priority
+
+  action {
+    type             = "forward"
+    target_group_arn = aws_lb_target_group.green[0].arn
+  }
+
+  condition {
+    path_pattern {
+      values = var.alb_paths == [] ? ["*"] : var.alb_paths
+    }
+  }
+
+  dynamic "condition" {
+    for_each = var.alb_host_header == null ? [] : [true]
+    content {
+      host_header {
+        values = [var.alb_host_header]
+      }
+    }
+  }
+
+  dynamic "condition" {
+    for_each = var.custom_header_token == "" ? [] : [true]
+    content {
+      http_header {
+        http_header_name = "custom-header-token" # Match value within cloudfront module
+        values           = [var.custom_header_token]
+      }
+    }
+  }
+
+  dynamic "condition" {
+    for_each = var.green_header_value == null ? [] : [true]
+    content {
+      http_header {
+        http_header_name = "green-header-name"
+        values           = [var.green_header_value]
+      }
+    }
+  }
+
+  tags = local.tags
+}
+
 /* -------------------------------------------------------------------------- */
 /*                                   Secret                                   */
 /* -------------------------------------------------------------------------- */