+++ CHANGELOG.md
@@ -5,8 +5,13 @@
 ### Added
 
 - Support blue-green deployment
-    - var: is_enable_blue_green_deployment
-    - resource:  aws_lb_target_group.green
+    - var: 
+        - is_enable_blue_green_deployment
+        - green_header_value
+    - resource:  
+        - aws_lb_target_group.green
+        - aws_lb_listener_rule.green
+
 
 ## [v1.3.1] - 2025-04-09